pyproject.toml
[project]
name = "dominusprime"
dynamic = ["version"]
description = "DominusPrime is a **personal assistant** that runs in your own environment. It talks to you over multiple channels (DingTalk, Feishu, QQ, Discord, iMessage, WhatsApp, Telegram, etc.) and runs scheduled tasks according to your configuration. **What it can do is driven by Skills — the possibilities are open-ended.** Built-in skills include cron, PDF/Office handling, news digest, file reading, and more; you can add custom skills. All data and tasks run on your machine; no third-party hosting."
readme = "README.md"
requires-python = ">=3.10,<3.14"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "agentscope==1.0.17",
    "agentscope-runtime==1.1.1",
    "httpx>=0.27.0",
    "discord-py>=2.3",
    "dingtalk-stream>=0.24.3",
    "uvicorn>=0.40.0",
    "apscheduler>=3.11.2,<4",
    "playwright>=1.49.0",
    "questionary>=2.1.1",
    "mss>=9.0.0",
    "reme-ai==0.3.0.5",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "python-socks>=2.5.3",
    "onnxruntime<1.24",
    "lark-oapi>=1.5.3",
    "python-telegram-bot>=20.0",
    "twilio>=9.10.2",
    "duckduckgo-search>=4.0.0",
]

[tool.setuptools.dynamic]
version = {attr = "dominusprime.__version__.__version__"}

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
# Dockerfile copies console build output to src/dominusprime/console/ (no dist subdir).
"dominusprime" = [
    "console/**",
    "agents/md_files/**",
    "agents/skills/**",
    "tokenizer/**",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
dominusprime = "dominusprime.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.2.1",
]
local = [
    "huggingface_hub>=0.20.0",
]
llamacpp = [
    "dominusprime[local]",
    "llama-cpp-python>=0.3.0",
]
mlx = [
    "dominusprime[local]",
    "mlx-lm>=0.10.0",
]
ollama = [
    "ollama>=0.6.1",
]
multimodal = [
    "pillow>=10.0.0",          # Image processing
    "opencv-python>=4.8.0",    # Video processing
    "transformers>=4.30.0",    # For CLIP embeddings
    "torch>=2.0.0",            # PyTorch for neural models
    "numpy>=1.24.0",           # Array operations
    "faiss-cpu>=1.7.4",        # Vector similarity search (optional)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
